Leave saving and dating y
Desired spreadsheets are saved in names that are not dated and in workbooks that contain multiple other irrelevant tabs. This code opens the desired files using a formatted name depending on the day of the week. (reports are for previous business day). It then saves and dates all of the files.